+++ src/main/java/com/upstage/devup/auth/controller/LogoutController.java
@@ -2,7 +2,7 @@ package com.upstage.devup.auth.controller;
 
 import com.upstage.devup.auth.config.AuthenticatedUser;
 import com.upstage.devup.auth.domain.dto.LogoutResponseDto;
-import com.upstage.devup.auth.exception.UnauthenticatedException;
+import com.upstage.devup.global.exception.UnauthenticatedException;
 import com.upstage.devup.global.provider.CookieProvider;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
