+++ src/main/java/com/upstage/devup/auth/service/UserAuthService.java
@@ -4,7 +4,7 @@ import com.upstage.devup.auth.config.jwt.JwtTokenProvider;
 import com.upstage.devup.auth.domain.dto.*;
 import com.upstage.devup.global.entity.User;
 import com.upstage.devup.auth.domain.mapper.UserMapper;
-import com.upstage.devup.auth.exception.InvalidLoginException;
+import com.upstage.devup.global.exception.InvalidLoginException;
 import com.upstage.devup.auth.respository.UserAuthRepository;
 import com.upstage.devup.global.exception.EntityNotFoundException;
 import com.upstage.devup.global.exception.ValueAlreadyInUseException;
