+++ src/main/java/com/upstage/devup/global/exception/InvalidLoginException.java
@@ -1,4 +1,4 @@
-package com.upstage.devup.auth.exception;
+package com.upstage.devup.global.exception;
 
 public class InvalidLoginException extends RuntimeException {
     public InvalidLoginException(String message) {
