+++ src/test/java/com/upstage/devup/user/statistics/service/UserSolvedQuestionReadTest.java
@@ -1,6 +1,6 @@
 package com.upstage.devup.user.statistics.service;
 
-import com.upstage.devup.auth.exception.UnauthenticatedException;
+import com.upstage.devup.global.exception.UnauthenticatedException;
 import com.upstage.devup.user.statistics.domain.dto.UserSolvedQuestionDto;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
